analytics/requirements.txt
attrdict==2.0.0
aiounittest==1.1.0
numpy==1.14.5
pandas==0.20.3
pyzmq==18.0.1
scipy==1.1.0
websockets==8.1